Read all "16" buttons from the G25/G27, two are totally unused. Apart from the 12 physical buttons one is for reverse detection and one is for sequential detection

+++ Joystick.h
@@ -60,7 +60,7 @@
 			int8_t  X; /**< Current absolute joystick X position, as a signed 8-bit integer */
 			int8_t  Y; /**< Current absolute joystick Y position, as a signed 8-bit integer */
 			int8_t  Z; /**< Current absolute joystick Z position, as a signed 8-bit integer */
-			uint8_t Button; /**< Bit mask of the currently pressed joystick buttons */
+			uint16_t Button; /**< Bit mask of the currently pressed joystick buttons */
 		} USB_JoystickReport_Data_t;
 
 	/* Macros: */
@@ -97,4 +97,3 @@
 		                                          const uint16_t ReportSize);
 
 #endif
-
